+++ aclpy/client/client_pkcs12.py
@@ -17,10 +17,13 @@ class ArrowheadClient(ArrowheadClientBase):
     Additional attributes:
     p12file (str) -- path to the .p12 certificate
     p12pass (str) -- password to the .p12 certificate
-    pubfile (str) -- path to the public key .pub
+    pubkey (str) -- public key, mutually exclusive with 'pubfile'
+    pubfile (str) -- path to the public key .pub, mutually exclusive with 'pubkey'
     cafile (str) -- path to the certificate authority file .ca
     server (ArrowheadServer) -- configuration of the Arrowhead Core server
     interfaces (List[ArrowheadInterfaces]) -- list of available interfaces, [] by default
+
+    Note: When pub* are not given, the public key is obtained from p12 file.
     """
 
     def __init__(self, *,
@@ -29,19 +32,36 @@ class ArrowheadClient(ArrowheadClientBase):
             port: int,
             p12file: str,
             p12pass: str,
-            pubfile: str,
+            pubkey: str = None,
+            pubfile: str = None,
             cafile: str,
             server: ArrowheadServer,
             interfaces: List[ArrowheadInterface] = [],
     ):
         """Initialize ArrowheadClient class."""
-        # Read pubkey first
-        with open(pubfile, "r") as f:
-            pubkey = f.read()
+        if pubkey is not None and pubfile is not None:
+            raise ValueError("Conflict betwen pubkey and pubfile. Provide only one of them.")
+
+        if pubfile is not None:
+            # Read pubkey first
+            with open(pubfile, "r") as f:
+                pubkey = f.read()
+
+        if pubkey is None:
+            from OpenSSL import crypto
+
+            pubkey = crypto.dump_publickey(
+                crypto.FILETYPE_PEM,
+                crypto.load_pkcs12(
+                    open(p12file, "rb").read(),
+                    p12pass
+                ).get_certificate().get_pubkey()
+            )
+
 
         self.connector = ArrowheadConnector(server)
 
-        super(ArrowheadClient, self).__init__(name, address, port, pubkey.replace("\n", ""), self.connector)
+        super(ArrowheadClient, self).__init__(name, address, port, str(pubkey).replace("\n", ""), self.connector)
 
         self.p12file = p12file
         self.p12pass = p12pass